small refactorings, aiming for an empty db_dive
