+++ octodeco/flaskr/dive.py
@@ -227,7 +227,7 @@ def get_cached_dive(dive_id: int):
     if cdp is None or cdp.profile_base() is None:
         session[ 'last_dive_id' ] = None;
         abort(404);
-    if not db_dive.is_display_allowed(cdp.profile_base()):
+    if not user.is_display_allowed(cdp.profile_base()):
         session[ 'last_dive_id' ] = None;
         abort(403);
     if not cdp.profile_base().is_ephemeral:
